third_party/blink/web_tests/external/wpt/web-animations/responsive/minHeight-expected.txt

This is a testharness.js-based test.
[PASS] minHeight responsive to style changes
[FAIL] minHeight clamped to 0px on keyframes
  assert_equals: expected "20px" but got "0px"
Harness: the test ran to completion.
